requirements/base.txt
Django==5.2.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
psycopg==3.3.2
